# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/babies/{babyId}/members": {
      "get": {
        "operationId": "FamiliesController_listMembers",
        "parameters": [
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/babies/{babyId}/invites": {
      "post": {
        "operationId": "FamiliesController_createInvite",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFamilyInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedFamilyInviteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      },
      "get": {
        "operationId": "FamiliesController_listInvites",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "accepted",
                "revoked",
                "expired"
              ],
              "type": "string"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyInviteListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/invites/preview": {
      "post": {
        "operationId": "FamiliesController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitePreviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/invites/accept": {
      "post": {
        "operationId": "FamiliesController_accept",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedInviteSuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/babies/{babyId}/invites/{inviteId}": {
      "delete": {
        "operationId": "FamiliesController_revoke",
        "parameters": [
          {
            "name": "inviteId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/babies/{babyId}/members/{memberId}": {
      "patch": {
        "operationId": "FamiliesController_updateMember",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFamilyMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyMemberResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      },
      "delete": {
        "operationId": "FamiliesController_removeMember",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
    "/api/v1/babies/{babyId}/membership": {
      "delete": {
        "operationId": "FamiliesController_leaveFamily",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "babyId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "families"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baby MP API",
    "description": "Baby MP HTTP API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "REFRESH_TOKEN_INVALID",
              "FORBIDDEN",
              "RESOURCE_NOT_FOUND",
              "VALIDATION_FAILED",
              "VERSION_CONFLICT",
              "IDEMPOTENCY_CONFLICT",
              "CONFLICT",
              "LAST_ADMIN_REQUIRED",
              "INVITE_INVALID",
              "INVITE_EXPIRED",
              "INVITE_REVOKED",
              "INVITE_ALREADY_USED",
              "ALREADY_A_MEMBER",
              "EXPORT_NOT_READY",
              "EXPORT_EXPIRED",
              "RATE_LIMITED",
              "UPLOAD_INCOMPLETE",
              "UPLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "INTERNAL_ERROR"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "example": "提交内容有误"
          },
          "requestId": {
            "type": "string",
            "example": "req_0198f76a-cd42-7000-8000-000000000001"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "FamilyUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "format": "uri"
          }
        },
        "required": [
          "id",
          "displayName",
          "avatarUrl"
        ]
      },
      "FamilyMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/FamilyUserSummaryDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "removed"
            ]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "number",
            "minimum": 1
          },
          "isCurrentUser": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "user",
          "role",
          "status",
          "joinedAt",
          "version",
          "isCurrentUser"
        ]
      },
      "FamilyMemberListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyMemberDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateFamilyInviteDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "editor",
              "viewer"
            ]
          },
          "expiresInHours": {
            "type": "number",
            "default": 24,
            "minimum": 1,
            "maximum": 168
          }
        },
        "required": [
          "role"
        ]
      },
      "CreatedFamilyInviteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "editor",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "revoked",
              "expired"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "inviter": {
            "$ref": "#/components/schemas/FamilyUserSummaryDto"
          },
          "token": {
            "type": "string",
            "description": "Returned only as the result of the idempotent create operation."
          },
          "sharePath": {
            "type": "string",
            "example": "/pages/family/invite?token=..."
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "expiresAt",
          "createdAt",
          "inviter",
          "token",
          "sharePath"
        ]
      },
      "CreatedFamilyInviteResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CreatedFamilyInviteDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "FamilyInviteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "editor",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "revoked",
              "expired"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "inviter": {
            "$ref": "#/components/schemas/FamilyUserSummaryDto"
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "expiresAt",
          "createdAt",
          "inviter"
        ]
      },
      "FamilyInviteListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyInviteDto"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "InviteTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Raw one-time invite token. It is accepted only in the JSON request body."
          }
        },
        "required": [
          "token"
        ]
      },
...
      "InvitePreviewDto": {
        "type": "object",
        "properties": {
          "baby": {
            "$ref": "#/components/schemas/InvitePreviewBabyDto"
          },
          "inviter": {
            "$ref": "#/components/schemas/FamilyUserSummaryDto"
          },
          "role": {
            "type": "string",
            "enum": [
              "editor",
              "viewer"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "revoked",
              "expired"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "baby",
          "inviter",
          "role",
          "status",
          "expiresAt"
        ]
      },
      "InvitePreviewResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InvitePreviewDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "AcceptedInviteResponseDto": {
        "type": "object",
        "properties": {
          "baby": {
            "$ref": "#/components/schemas/BabyDto"
          },
          "member": {
            "$ref": "#/components/schemas/FamilyMemberDto"
          }
        },
        "required": [
          "baby",
          "member"
        ]
      },
      "AcceptedInviteSuccessResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AcceptedInviteResponseDto"
          }
        },
        "required": [
          "data"
        ]
      },
      "UpdateFamilyMemberDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "minimum": 1
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ]
          }
        },
        "required": [
          "version",
          "role"
        ]
      },
      "FamilyMemberResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/FamilyMemberDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
    }
  }
}